% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{epitheme_gt}
\alias{epitheme_gt}
\title{Table styling for gt table}
\usage{
epitheme_gt(table, convert_gt = TRUE)
}
\arguments{
\item{table}{a table to be converted to gt}

\item{convert_gt}{boolean. Is there a need convert the df to gt ?}
}
\value{
a gt object formatted
}
\description{
Table styling for gt table
}
\examples{

tab <- iris|>
dplyr::summarise(
  .by = Species,
  n = dplyr::n(),
  max_sepal = max(Sepal.Length, na.rm = TRUE),
  n_setosa = sum(Species == "setosa", na.rm = TRUE),
  setosa_pct = round(digits = 2, (n_setosa / n) * 100)
)

tab |>
  epitheme_gt() |>
  gt::tab_footnote("This dataset is pretty boring")

}
