% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epithemes-package.R
\docType{package}
\name{epithemes-package}
\alias{epithemes}
\alias{epithemes-package}
\title{epithemes: Epicentre (MSF) themes and colors}
\description{
provides function to generate colour palettes, ggplot themes and gt style for Epicentre.
}
\keyword{internal}
