% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{get_hex}
\alias{get_hex}
\title{Extract Hex value from a palette}
\usage{
get_hex(palette, color_name)
}
\arguments{
\item{palette}{The name of the palette}

\item{color_name}{The name of the color to extract}
}
\value{
a length 1 names vector of color_name and the according hex value
}
\description{
Extract Hex value from a palette
}
\examples{
epithemes::get_hex("main", "primary")
}
