% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{print_palette}
\alias{print_palette}
\title{Print a palette}
\usage{
print_palette(palette, n)
}
\arguments{
\item{palette}{The name of the chosen palette. Choices are from: \code{"epicentre-main"}.}

\item{n}{The integer length of the desired palette. Default value is the length of the specified palette.}
}
\value{
A plot showing the colours of the chosen colour palette.
}
\description{
Prints the chosen palette of length \code{n}.
}
\examples{
print_palette(palette = "epicentre-main")
print_palette(palette = "epicentre-main", n = 5)
print_palette(palette = "epicentre-main", n = 5)
}
