% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{scale_fill_epi}
\alias{scale_fill_epi}
\title{Make a ggplot fill scale}
\usage{
scale_fill_epi(palette, n)
}
\arguments{
\item{palette}{The name of the chosen palette. Choices are from: \code{"main"}.}

\item{n}{The integer length of the desired palette. Default value is the length of the specified palette.}
}
\value{
A discrete scale to use for fill in ggplot.
}
\description{
Makes a discrete fill scale of length \code{n} from the chosen palette.
}
\examples{
library(ggplot2)

ggplot(data = iris,
aes(x = Species, y = Petal.Width, fill = Species)) +
geom_violin() +
epithemes::scale_fill_epi(palette = "main", n = 3)

}
