% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{view_logo}
\alias{view_logo}
\title{Visualise a logo}
\usage{
view_logo(logo_name)
}
\arguments{
\item{logo_name}{Name of a logo. Available logos are found with list_logo()}
}
\value{
plots a rastergrob grob of given logo
}
\description{
Visualise a logo
}
