% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{epitheme_gt}
\alias{epitheme_gt}
\title{Table styling for gt table}
\usage{
epitheme_gt(
  table,
  convert_gt = TRUE,
  source_note = NULL,
  fmt_percent = NULL,
  pct_decimal = 1,
  fmt_na = "*-*",
  stripped = TRUE
)
}
\arguments{
\item{table}{a table}

\item{convert_gt}{`lgl` Convert the `table` to gt ?. Default is `TRUE`}

\item{source_note}{`chr` footnote to add to the table}

\item{fmt_percent}{`var` variables to format to percentages}

\item{pct_decimal}{`num` decimals for percentages}

\item{fmt_na}{`chr` Markdown string to format `NA` values}

\item{stripped}{`lgl` Show stripped rows?}
}
\value{
a gt object formatted
}
\description{
Table styling for gt table
}
\examples{

iris|>

dplyr::summarise(
  .by = Species,
  n = dplyr::n(),
  max_sepal = max(Sepal.Length, na.rm = TRUE),
  n_setosa = sum(Species == "setosa", na.rm = TRUE),
  setosa_pct = round(digits = 2, (n_setosa / n) * 100)
) |>

  epitheme_gt() |>

  gt::tab_footnote("This dataset is pretty boring")

}
